chromium/third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/subtree-visibility-052.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Subtree Visibility: accessibility press</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<link rel="match" href="container-ref.html">
<meta name="assert" content="subtree-visibility auto subtrees are exposed by accessibility press">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.auto {
  subtree-visibility: auto;
}
</style>

<div id="hidden" style="subtree-visibility: auto">
  foo
  <div id="child" tabindex="0">
    bar
  </div>
</div>

<script>
function axElementById(id) {
  return accessibilityController.accessibleElementById(id);
}

async_test(async(t) => {
  const hiddenEl = document.getElementById("hidden");
  let axHidden = axElementById("hidden");
  t.step(() => { assert_equals(axHidden.childrenCount, 3, "Child count after acquire"); });

  axElementById("child").press();

  // Wait for the next frame for the ax object to be recreated.
  requestAnimationFrame(() => {
    requestAnimationFrame(() => {
      axHidden = axElementById("hidden");
      t.step(() => { assert_equals(axHidden.childrenCount, 2, "Child count after activation"); });
      t.done();
    });
  });
}, "Accessiblility press() causes activatable hidden tree to activate");

</script>
